[NFC] Move tensor.pad related passes to Common/
