+++ compiler/src/iree/compiler/Codegen/Common/BUILD
@@ -101,16 +101,19 @@ iree_compiler_cc_library(
         "BufferizationAnalysis.cpp",
         "BufferizeCopyOnlyDispatchesPass.cpp",
         "CleanupBufferAllocViewPass.cpp",
+        "ConcretizePadResultShape.cpp",
         "ConvertToDestinationPassingStylePass.cpp",
         "FlattenMemRefSubspanPass.cpp",
         "FoldTensorExtractOpPass.cpp",
         "ForOpCanonicalizationPass.cpp",
+        "FuseTensorPadWithConsumer.cpp",
         "IREEComprehensiveBufferizePass.cpp",
         "OptimizeVectorTransferPass.cpp",
         "PolynomialApproximationPass.cpp",
         "TestPartitionableLoopsInterface.cpp",
         "TileAndDistributeToWorkgroupsPass.cpp",
         "TypePropagationPass.cpp",
+        "VectorizePad.cpp",
     ],
     hdrs = [
         "BufferizationAnalysis.h",
