+++ compiler/src/iree/compiler/Codegen/SPIRV/CMakeLists.txt
@@ -27,7 +27,6 @@ iree_cc_library(
     "Passes.cpp"
     "SPIRVCreateFastSlowPath.cpp"
     "SPIRVDistribute.cpp"
-    "SPIRVFuseTensorPadWithConsumer.cpp"
     "SPIRVLowerExecutableTargetPass.cpp"
     "SPIRVTile.cpp"
     "SPIRVTileAndDistribute.cpp"
@@ -36,7 +35,6 @@ iree_cc_library(
     "SPIRVVectorToCooperativeOps.cpp"
     "SPIRVVectorize.cpp"
     "SPIRVVectorizeLoadStore.cpp"
-    "SPIRVVectorizePad.cpp"
     "Utils.cpp"
   DEPS
     IREELinalgExtDialect
