+++ compiler/src/iree/compiler/Codegen/SPIRV/SPIRVTile.cpp
@@ -57,106 +57,6 @@ static void populateTilingReductionPatterns(RewritePatternSet &patterns) {
                                                    filter);
 }
 
-/// Gets the given `attrOrValue` as an index value by creating constant ops
-/// for attributes.
-static Value getAsIndexValue(OpFoldResult attrOrValue, OpBuilder &builder,
-                             Location loc) {
-  IntegerAttr attr;
-  if (Value val = attrOrValue.dyn_cast<Value>()) {
-    if (val.getType().isIndex()) return val;
-    matchPattern(val, m_Constant(&attr));
-  } else {
-    attr = attrOrValue.get<Attribute>().cast<IntegerAttr>();
-  }
-  return builder.createOrFold<arith::ConstantIndexOp>(
-      loc, attr.getValue().getSExtValue());
-}
-
-namespace {
-
-/// Concretizes tensor.pad op's result shape if its source op implements
-/// OffsetSizeAndStrideOpInterface. For example, pad(extract_slice).
-struct ConcretizePadResultShape final : public OpRewritePattern<tensor::PadOp> {
-  using OpRewritePattern::OpRewritePattern;
-
-  LogicalResult matchAndRewrite(tensor::PadOp padOp,
-                                PatternRewriter &rewriter) const override {
-    // If the result shape is already static, then nothing to do.
-    if (padOp.getResultType().hasStaticShape()) return failure();
-
-    int rank = padOp.getResultType().getRank();
-    SmallVector<int64_t> staticShape;
-    staticShape.reserve(rank);
-
-    auto sourceIfxOp = dyn_cast_or_null<OffsetSizeAndStrideOpInterface>(
-        padOp.getSource().getDefiningOp());
-    if (!sourceIfxOp) return failure();
-
-    SmallVector<OpFoldResult> lowPad = padOp.getMixedLowPad();
-    SmallVector<OpFoldResult> source = sourceIfxOp.getMixedSizes();
-    SmallVector<OpFoldResult> highPad = padOp.getMixedHighPad();
-
-    MLIRContext *context = padOp.getContext();
-    Location loc = padOp.getLoc();
-
-    AffineExpr sym0, sym1, sym2;
-    bindSymbols(context, sym0, sym1, sym2);
-    auto addMap = AffineMap::get(0, 3, {sym0 + sym1 + sym2}, context);
-
-    SmallVector<Value, 3> valueSizes;
-    for (int dimIndex = 0; dimIndex < rank; ++dimIndex) {
-      valueSizes.clear();
-      valueSizes.push_back(getAsIndexValue(lowPad[dimIndex], rewriter, loc));
-      valueSizes.push_back(getAsIndexValue(source[dimIndex], rewriter, loc));
-      valueSizes.push_back(getAsIndexValue(highPad[dimIndex], rewriter, loc));
-
-      // The pad op's result shape is low padding + source size + high padding.
-      // Try to see if we can get a constant number by composing and
-      // canonicalizing the result. We use affine mechanisms here because
-      // generating arithmetic add ops over dim ops won't work, given they are
-      // SSA values that would need invoking other patterns to simplify. We
-      // cannot invoke patterns in patterns.
-      AffineMap map = addMap;
-      fullyComposeAffineMapAndOperands(&map, &valueSizes);
-      canonicalizeMapAndOperands(&map, &valueSizes);
-
-      auto cstExpr = map.getResult(0).dyn_cast<AffineConstantExpr>();
-      // Specially handle the case where we have both dimensions and symbols and
-      // they map to the same value, e.g.:
-      //   affine_map<(d0, s0) -> (d0 - s0 + 4)>(%v, %v).
-      // Due to the restrictions over dimensions and symbols, the above won't
-      // simplify. Try to change dimensions for symbols for such cases.
-      if (!cstExpr && llvm::all_equal(valueSizes)) {
-        int numDims = map.getNumDims();
-        int numSyms = map.getNumSymbols();
-        DenseMap<AffineExpr, AffineExpr> dimToSymMap;
-        for (int i = 0; i < numDims; ++i) {
-          dimToSymMap[rewriter.getAffineDimExpr(i)] =
-              rewriter.getAffineSymbolExpr(numSyms + i);
-        }
-        map = map.replace(dimToSymMap, /*numResultDims=*/0,
-                          /*numResultSyms=*/numDims + numSyms);
-
-        canonicalizeMapAndOperands(&map, &valueSizes);
-        cstExpr = map.getResult(0).dyn_cast<AffineConstantExpr>();
-      }
-      if (!cstExpr) return failure();
-
-      staticShape.push_back(cstExpr.getValue());
-    }
-
-    auto resultType = RankedTensorType::get(
-        staticShape, padOp.getResultType().getElementType(),
-        padOp.getResultType().getEncoding());
-
-    rewriter.updateRootInPlace(
-        padOp, [&]() { padOp.getResult().setType(resultType); });
-    return success();
-  }
-};
-
-}  // namespace
-
 //===----------------------------------------------------------------------===//
 // Main pass
 //===----------------------------------------------------------------------===//
@@ -259,20 +159,9 @@ class SPIRVTilePass final : public SPIRVTileBase<SPIRVTilePass> {
       });
     }
 
-    {  // Canonicalize.
-      RewritePatternSet patterns =
-          linalg::getLinalgTilingCanonicalizationPatterns(context);
-      // Pulling in upstream scf.for and affine.min canonicalization patterns.
-      // They work on tiled (but not distributed) loops.
-      scf::populateSCFForLoopCanonicalizationPatterns(patterns);
-      // Pulling in IREE scf.for and affine.min canonicalization patterns.
-      // They work on tiled and distributed loops.
-      populateFoldAffineMinInDistributedLoopsPatterns(patterns);
-      // Pulling in flow.dispatch.tensor.load op canonicalization patterns.
-      // Tiling can generate dim ops taking them as operands.
-      IREE::Flow::DispatchTensorLoadOp::getCanonicalizationPatterns(patterns,
-                                                                    context);
-      patterns.add<ConcretizePadResultShape>(context);
+    {
+      RewritePatternSet patterns(context);
+      populateConcretizePadResultShapePatterns(context, patterns);
       (void)applyPatternsAndFoldGreedily(funcOp, std::move(patterns));
 
       LLVM_DEBUG({
@@ -323,22 +212,13 @@ class SPIRVTilePass final : public SPIRVTileBase<SPIRVTilePass> {
       });
     }
 
-    {  // Canonicalize.
-      RewritePatternSet patterns =
-          linalg::getLinalgTilingCanonicalizationPatterns(context);
-      // Pulling in upstream scf.for and affine.min canonicalization patterns.
-      // They work on tiled (but not distributed) loops. We only tiled reduction
-      // loops previously so this should be fine.
-      scf::populateSCFForLoopCanonicalizationPatterns(patterns);
-      // Pulling in flow.dispatch.tensor.load op canonicalization patterns.
-      // Tiling can generate dim ops taking them as operands.
-      IREE::Flow::DispatchTensorLoadOp::getCanonicalizationPatterns(patterns,
-                                                                    context);
-      patterns.add<ConcretizePadResultShape>(context);
+    {
+      RewritePatternSet patterns(context);
+      populateConcretizePadResultShapePatterns(context, patterns);
       (void)applyPatternsAndFoldGreedily(funcOp, std::move(patterns));
 
       LLVM_DEBUG({
-        llvm::dbgs() << "--- After tiling canonicalization  ---\n";
+        llvm::dbgs() << "--- After tiling canonicalization ---\n";
         funcOp.print(llvm::dbgs(), OpPrintingFlags().useLocalScope());
         llvm::dbgs() << "\n\n";
       });
